Add fields check to physical property models (#1065)
Description of the problem
Some physical and interface property models require certain fields to compute property values and partial derivatives. When these fields were not defined a not so debug-friendly message was printed out.
Description of the solution
To help the debugging process, if an error were to occur, assertions were added to the property models in functions where the fields were required.
How Has This Been Tested?
By commenting lines in unit tests and code, the throw of the exception was verified in DEBUG mode.
Comments
Doxygen descriptions were also added to specific_heat_model.h.