Use the eigenvalue estimation of PreconditionRelaxation (#1064)
Description of the problem
The eigenvalue estimation for the smoother of the multigrid preconditioners was done through a PreconditionChebyshev.

Description of the solution
Due to recent changes in deal.II (dealii/dealii#16742), now we are able to do it using the PreconditionRelaxation directly. This allows us to remove the estimate omega function since it is done internally by the smoother if the relaxation parameter is set to 1. This allows also to remove the degree parameter that was only used in the case of Chebyshev.