Make stasis constraint compatible with restart and adaptive mesh refi… (#1072)
Description of the problem
The recently implemented vof stasis constraint (Implement temperature-dependant solid domain contraints in VOF #1048) was not compatible with adaptive mesh refinements and checkpointing system. When refining the mesh or when restarting a simulation, the filtered phase fraction was reinitialized, but not filled with the appropriate values.

Description of the solution
The apply_phase_filter() function was added to post_mesh_adaptation() and read_checkpoint() functions.

How Has This Been Tested?
The stasis constraint was tested using the 2D LPBF benchmark :
with restart and mesh adaptation;
without restart and with mesh adaptation, and;
with restart and without mesh adaptation.

Former-commit-id: bb7cea36fdf4d2750c449399da03587f57b02386