Dem volume insertion box parameters (#1074)
Description of the problem
The insertion box for the volume insertion was being defined by a large number of parameters (9 to be exact)

Description of the solution
The box is defined by one parameter and the order of direction by an other.

Documentation
The documentation has been updated accordingly.

Comments
The list insertion parameters are now parsed with "convert_string_to_vector", which is cleaner to look at.
I have made (with the help of Amishga) a bash script to parse every prm file and change the old parameters into the new ones. We could do something with it... maybe?