Dem insertion naming refactor (#1079)
Description of the problem
The naming scheme for the insertion files were not coherent with the rest of the code. (i.e., integrator_... , particle_wall_... , etc.). This was annoying when trying to find a specific insertion file. Now they are next to each other.
Description of the solution
The files are now named like this: insertion_method.cc or insertion_method.h
How Has This Been Tested?
Tests are running on my machine with these changes.