Fix restart transient with variable time step (#1093)
Description of the problem
Transient simulation with adaptive time-stepping were not restarted correctly (their time-step was not fixed adequately)
Description of the solution
The issue was that the time-step value was not reset correctly when the time-step values vector was read in the checkpoint
How Has This Been Tested?
Added. a new test for this feature
Documentation
Changelog added
Comments
This class uses an old checkpointing mechanism that is not boost serialize. I will need to fix it. It will be a good opportunity to streamline the class afterwards also and make it simpler to understand. Maybe reduce the tree of possibilities in term of classes.

Co-authored-by: Pierre Laurentin <81017102+PierreLaurentinCS@users.noreply.github.com>