Flow rates postprocessing (#1092)
Description of the problem
The previous flow rate post processing tool would add a line to the .dat file for each defined boundary (in the boundary condition subsection of the .prm file) for every time iteration. This is not the expected output as one would expect a single line to be added for each time step with the flow rates of every boundary.
Also, the values of the flow rates were added with a 12 digits precision. Instead, a 4 digits precision in scientific notation was chosen, which seems like a reasonable precision.
Description of the solution
The value of time is now added outside of the loop on the boundary conditions.
Comments
At the moment there is no unified output format : with/without scientific notation, different numbers of digits. This could be the work of a future PR.