Fix nitsche restart (#1106)
Description of the problem
Nitsche does not restart when the mesh was generated with simplices. See Nitsche model restart does not work with simplices #275
Description of the solution
Fixes the issue by restarting the fully distributed triangulation for the solid correctly.
How Has This Been Tested?
Still pending a unit test, WIP
Comments
While running this test case I managed to find another bug, which is that the Nitsche model is unable of mesh adaptation once it has been restarted. This bug was there all the time. I will try to figure out something, in the meantime I will try to m erge this PR anyway.

Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>