Improve Scratch data for CFD-DEM simulations (#1115)
Description of the problem
The scratch data for CFD-DEM simulation was very much all over the place. Memory was being reallocated and interpolation was sometimes done manually whereas some other times it was done using a quadrature defined on the fly
Description of the solution
Unify interpolation
Prevent memory reallocation
How Has This Been Tested?
All tests pass with no changed results (as expected)
Comments
This does not lead to a large speed boost (like 3-4% on what I tested), but it really makes the code more readable and more uniform. Next PR will be on improving the assemblers. Since this was really limited to the scratch data I feel this PR can stand on its own.