Add more coarse-grid solvers for GMG  (#1121)
Description of the problem
Until now the matrix-free solver was only supporting GMRES as coarse-grid solver preconditioned by ILU or AMG for the GMG preconditioner.

Description of the solution
This PR adds also the possibility to choose AMG, ILU or a direct solver as coarse-grid solver in the GMG preconditioner. It creates a new parameter that allows to choose the coarse-grid solver and renames the GMRES specific parameters to be able to differentiate them.

How Has This Been Tested?
All the tests use by default the GMRES coarse-grid solver and none of them was modified. Only the parameters were renamed according to the new definition of GMRES parameters in the prm file.

Documentation
 doc/source/parameters/cfd/linear_solver_control.rst: the parameters were updated in the linear solver section.
 doc/source/examples/incompressible-flow/3d-taylor-green-vortex/3d-taylor-green-vortex.rst: the name of the gmres parameters was updated.