Add advection to the adaptive sparse contacts (ASC) (#1113)
Description of the problem
Periodic nodes were not properly handled with the last implementation: the periodic nodes did not have information on the mobility status on the coincident nodes. This was spotted when there has particles at one side of the periodic cells and the other periodic ones were empty. Those cells where not mobile, but this is one of the criteria for mobile.

Disabling contact of particles was functional in CFD-DEM: cells with low granular temperature deactivated the contacts. However, it is possible for particles within a cell to have a low variance in velocity and a significant average velocity, such as when particles experience bulk motion.

Description of the solution
Periodic nodes:

Mapping periodic nodes was achieved with the constraints since we can identify the periodic nodes from it (so called identity constraints).
Some attempted as been done to apply the periodic constraints to the mobility_at_nodes, without any success. The main issue is that constraints force values at PB1 = values at PB2. In our case, we need to check of the prevailing value, so sometime it is PB2 = PB1, with was not compatible with straight constraints operations. "[...] it adds the DoFs of face_1 to the list of constrained DoFs in constraints and adds entries to constrain them to the corresponding values of the DoFs on face_2. " see Deal.II doc here
Disabling contacts with advection:

New mobility status has been added to the disabling contacts: advected and advected_active
The new algorithm doesn't detect when particles are advected, this is a user-defined parameter. It could be useful for DEM when particles are moving in bulk by gravity, but this is more practical for CFD-DEM. The first idea was to implement advected particles auto-detection, but this results in more checks, also the average velocities threshold is not easly configurable.
The way to "apply advection" is that the average velocity and acceleration of the cells are calculated or updated everytime the mobile cells are integrated. The average values are calculated in the same loop than the integration and it allows to always have the last calculated velocity values. For advected cells, all particles of a cell are integrated with an cell-averaged velocity and acceleration * time step. This velocity is also updated with the a*dt values.
How Has This Been Tested?
DEM tests for ASC was been updated
New test in CFD-DEM called adaptative_sparse_contacts.prm
Documentation
Change the "disabling contacts" parameters to "adaptive sparse contacts"
Doc is updated with new parameters and new names
Comments
This algorithm was called Disabling Contacts when it was first introduce in the code for DEM
An example will be added in another PR