Rename source term NS subsection (#1130)
Description of the problem
The subsection's name to specify the source term of the Navier-Stokes equations in the parameter file was set to be xyz.

Description of the solution
The subsection was renamed to fluid dynamics, since there are source terms for other physics where the subsection has the physics name, for example, subsection heat transfer. The change had to be made in all the parameter files of the examples, the tests and the documentation.