Allow to use viscous initial conditions for transient problems (#1134)
Description of the problem
The viscous initial condition was already implemented but we were not able to use it for transient simulations.

Description of the solution
This was fixed by setting the assembly method to steady before solving the non linear system with the specified viscosity. This was already done for the ramp viscous initial condition.

How Has This Been Tested?
Run a test on my machine and works just as expected.