Bunny drill fix (#1131)
Description of the problem
The bunny drill example was broken since the change to Solid Object which were now restarted
Description of the solution
Fix the problem by adding the bunny from the onset of the problem. However, this required changing insertion mechanism. I used this opportunity to demonstrate the plane insertion as part of an example. I think it's a cool idea.
For a reason I don't know, the time-step was way too small. Updated it to a more coherent value, the simulation is much faster now (like 10 minutes or so)
How Has This Been Tested?
Example was re-run and identical results are obtained
Documentation
Documentation is updated accordingly since we change the insertion mechanism
Comments
Working on this I found another bug. Right now the example will not work correctly if there are no load balancing steps. It seems to be related to the way the objects are displaced and the displacement monitoring is made after the checkpointing. I will use this example to work on fixing this bug, but this will be work this week in a next PR.

Closes the issue in #1129

Co-authored-by: Audrey Collard-Daigneault <71884806+acdaigneault@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>