Solid objects did not trigger a contact detection step when they were moved  (#1136)
Description of the problem
Moving solid objects did not trigger a contaction step when they moved. This was problematic since if the particles were static, the solid objects could waltz through them if no contact detection were triggered on the particle side.
Description of the solution
Trigger a contact detection step when the solid object is mapped again.
How Has This Been Tested?
Bunny example without load balancing works fine now.
Comments
This may have a small negative impact on performance. This will improve when we refactor the contact detection mechanism in the code to make it cleaner and more logical.


Former-commit-id: ef765b6f53fe9b011545ee56a268f35aa8f4b622