Robin/Neumann BC parameters were not space dependent (#1128)
Description of the problem
The convection-radiation-flux boundary condition had function expression parameters that were not space dependent.
Description of the solution
Made it space-dependent using the function evaluation at the quadrature point position.
How Has This Been Tested?
Simple 2D heat flux case. The test was added to the application tests.
Documentation
Remove disclaimer from parameters guide regarding the no longer existent limitation.
Comments
This PR is a follow up of PR Neumann BC for heat #1000 .
There is no way to have set number of fluids = 0. This was added to the maintenance project and I will fix it in another PR.

Co-authored-by: voferreira <voferreira@outlook.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>