Modify viscosity in stabilization tau (#1126)
Description of the problem
For phase change problem, the viscosity used in the tau computation for the stabilization in the solid used the numerical solid viscosity. Since this solid viscosity has no physical meaning, it was weird to used it, and the linear solver seemed to converge slower.
Description of the solution
Introduce new attributes in NS scratch data kinematic_viscosity_for_stabilization and dynamic_viscosity_for_stabilization. Both are computed by the rheology model. For now, except in the phase change, it returns the current viscosity. However, in the phase change rheology model, it returns the liquid viscosity.
How Has This Been Tested?
The melting cavity case was used to insure that the results stay the same and to show better convergence of the linear solver.

 heat_transfer_phase_change_constrain_solid_domain

 heat_transfer_vof_phase_change_constrain_solid_domain

Those two tests were changed to have a more representative behaviour.

Future changes
Test the viscosity scale in VOF and non-newtonian cases to see if it brings some gains

Co-authored-by: Pierre Laurentin <81017102+PierreLaurentinCS@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>

Former-commit-id: 542068e75e650705bf4c4a56fe43099e417c8b09