Fix dsp multiphysics (#1151)
Description of the problem
Several tests were crashing due to a new assert introduced in deal.II recently #dealii/dealii#17033. The dynamic sparsity pattern in most of the multiphysics was being initialized with the number of dofs.

Description of the solution
The dynamic sparsity pattern is now initialized using the locally relevant dofs as it should have been done since the beginning. The CI is not broken anymore.

How Has This Been Tested?
All test pass.

Former-commit-id: 1071a5e04b8f81cd079e0842277ffc8fabd7373f