Dem insertion: clear and add (#1135)
Description of the problem
LPBF spreading simulations requires a lots of particles. Most of them are useless during most of the simulation time, as they are deep the powder dispenser. It would be more efficient to only simulate the particles required for the layer N and not for the following layers.
Description of the solution
A new insertion method called "clear and add" has been implemented.
This method is very similar to file insertion in that it reads particle position, diameter and velocity from input files.
First, we define a clearing box, where particles will be deleted. Then, we insert new particles from an insertion file. Each insertion, a different insertion file can be use.
How Has This Been Tested?
Custom parameter file. (See slack)
Comments
I still need to do a test, and example and the documentation for this method.
I have change the order for the parsing of the parameters. It is now in alphabetical order. It is cleaner IMO.

Former-commit-id: f69c03c9033790a8b0b8ebac6e0c2ad446ec0058