Add refinement at frequency tests (#1163)
Description of the problem
The "refinement_at_frequency" member is set by setting the frequency to 0.
There was not test for this.
Description of the solution
2 tests were added.

How Has This Been Tested?
[application_tests/lethe-fluid/refinement_at_frequency_on] With frequency set to 1
[application_tests/lethe-fluid/refinement_at_frequency_on] With frequency set to 0

Comments
I would have preferred to have a true unit test for this instead of an application test, because with these changes the other physics are not tested. I'll keep looking for a way to test the parser itself.