Handle the deprecated parameters of individual entries to a list entry (#1171)
Description
Parameters are often changed as refactoring in the DEM solver, but it causes issues when running simulation with old parameter files.
Those files get incompatible with the latest version of the code and need to be updated by the user.
There is a function in deal.II (called declare_alias()) that allows to still use old parameters as alias, but need to be the same type so the parsing stays the same. Also, it can give a warning about the deprecated file.
We started, and we are planning, to convert some old individual parameter entries to one parameter that is a list of values.
It means 3 parameters of double will be changed to one parameter of a list of double.
In order to avoid the above, this refactoring aims to handles those old parameters.
The only issue with the current implementation is there is no way to assert if the new parameter is set as a double.
Also, this PR applies those changes to the gravity force in DEM.

Testing
New test deprecated_parameter uses the old gx, gy, gz parameters.
The old parameters (in 2d or 3d) were tested and it works well and gives a warning for the deprecated parameter.

Documentation
The parameters gx, gy and gz are now g.