Dem force chain ghost (#1172)
Description
Finalization of the force chains feature. Previously, the feature did not account for contact between local particles and ghost particles. This PR implements this calculattion the calculate_force_chains method. PVD output file is added in the write_force_chains method to ensure that multiple force chain time-steps can be visualized. 

Testing
Tested with the rectangular hopper example.

Documentation
The documentation has been is updated in the post-processing.rst
A new parameter in the post-processing part of parameter files is force chains.