Fix NS-VOF stong residual formulation (#1180)
Description
The term associated with the viscosity jump in the strong residual for the NS-VOF assembler added in PR Add missing VOF term in the strong form of the residual #1149 led to a ill-posed formulation in phase change cases where the solid is represented by a highly viscous fluid. Hence, we don't consider this term in the current formulation.
There was no appropriate application test for lpbf/phase change with vof cases that could have allowed us to identify this ill-posed formulation prior to the merge of PR Add missing VOF term in the strong form of the residual #1149.
The LPBF benchmark example prm files were almost up to date, but did not exactly matched the one we used to generate the benchmark results. The mesh files were missing from the repo.

Solution
This PR removes the term associated with the viscosity jump.
A new test is now included to avoid future mistake in lpbf/phase change with vof cases.
LPBF benchmark example prm files were modified and are now up to date with the current working version. The mesh files are now available in the repo.

Testing
The new test simulate the lpbf benchmark case using the extra-extra-coarse mesh. Temperature statistics and mass conservation monitoring are the outputted metrics (to track both the change in HT and NS-VOF solvers).
Other NS-VOF tests were updated, and results for those tests returned to their previous values (or almost with small differences).
Documentation
No modification in the documentation.

Co-authored-by: Amishga Alphonius <amishga1999@gmail.com>
Former-commit-id: b1ec9c1c4da37780400f9db5fb4eb3284bff9139