Optimize NS-VOF assembler (#1188)
Description
The NS-VOF assembler was not optimized as the NS assembler. The if conditions on the dof components were added to not compute multiplications that are a priori 0.

Testing
None of the application tests have changed. The computational time gain is about 20% by iteration on the 3D lpbf benchmark case for a simulation of about 300 iterations. I can do it for a complete simulation if you thing it is worth doing it :)

Documentation
No new documentation needed.