Nitsche checkpoint particles (#1192)
Description
In the previous version of the code, the lethe-fluid-nitsche solver would crash when trying to checkpoint a simulation that was already restarted. This was because the checkpointing mechanism was poorly implemented. Indeed, instead of checkpointing the particle_handler itself, the solver would actually re-load the triangulation of the solid object and generate new particles. This is not a good approach and is not maintanable.
Solution
I fought agaisnt my laziness and re-implemented the thing to actually checkpoint and load the particles. This now enables things to restart multiple times.
Testing
To test this feature, I updated the existing restart test so that it actually tries to keep on checkpointing. This would throw an assert in Debug. This does not happen anymore and things are all fine and dandy
Documentation
Nothing here needs to be documented, but some issues can be closed because of this.

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>
Former-commit-id: 5c7adf691a087f1fa168113a4ee434cf45bab96c