Add plane to stasis constraint and remove pressure constraints (#1193)
Description
Following the implementation of temperature-dependent solid domain constraint for both single-phase and two-phase flows, a need for an additional condition was identified. Indeed, in some simulations, splatters may happen, and with the previous implementation, splatter could freeze "mid-air" which is non-physical and therefore undesired. This PR adds a plane to divide the problem domain in two parts and only the one in the opposite direction to the normal vector is considered for stasis constraints to be applied. The plane is defined through a point and an outward pointing normal vector.

During this PR, a bug regarding the pressure constraints was detected and fixed. Previously, the pressure DOFs were also constrained, but this resulted in an ill-posed problem when sources terms in the momentum equation were pressure-dependent. Therefore, as of this PR, the constraints on pressure DOFs were removed.

Testing
Previous tests with solid domain constraints were updated due to the pressure DOFs that are no more constrained.
Additionally, two new application tests were added for the new plane restriction feature (1 for single fluid flows and one for VOF).

Documentation
The "Constrain Stasis" page in the parameter documentation was updated with the new features and disabled ones.



Former-commit-id: c41f675081b2e85147f46adce48d64c2bfaf1a7d