Upgrade reference format in documentation (DEM, CFD-DEM unresolved and RPT) (#1196)
Description
Previous referencing format was done by copy/pasting regular hyperlinks with manual increment [1], ­[2], ...
It was working fine, but was a bit redundant and error-prone.
reStructuredText has markup for citations, but also need manual increment if we want to associate a number to the citation.
I propose to use the footnote markup which is very similar to the citation markup, but allows to use labels that are not shown in the text.
This way it is easy to just add references info in the proper section and add the in-text reference to the text with the label. Clicking on the number will redirect to the reference section first and not the link of the article.

Documentation
There is a section in contributing.rst that explain the new formatting.