Encapsulation of the load balancing in DEM and CFD-DEM (#1199)
Description
The functions related to the load balancing were implemented in the DEM solver and coupling CFD-DEM solver even tho it is not really specific to the solvers.
The setup and the routine were duplicated in both solvers, making modifications or maintenances harder.
The solution is the implementation the load balancing class with mostly the same functions.
The load balancing is also fixed for the CFD-DEM. It was not working since nov. 2022, oops.

Testing
All tests related to the load balacing in DEM are still passing.
A new test spouted_bed_load_balancing_generator.prm was added for the unresolved cfd-dem tests.
After a quick fix on the master, results were compared and are the same.

Former-commit-id: 2e3dfc0609940735a816af1ab022b4d07cff8c0a