Rearrange the insertion documentation for DEM (#1202)
Description
There are now 4 insertion methods in DEM that some parameters are used by one method and some are used by many methods. @OGaboriault did a nice job by splitting the documentation into sections and specifiying what parameters are used with which methods. However, it seems still confusing from a novice's perspective because of the specificities on the parameters according to the method.
I introduce all the parameters that can be used for each method, so we can clearly know if I wanna use method X I can use parameter a, b, c, ...
It also closes issue #1198