Remove forcing particle displacement in PBC with CFD-DEM at last DEM iteration (#1204)
Description
When using PBCs in CFD_DEM, the last DEM iteration of a CFD iteration was forcing the execution of the displacement (translation the location of particles from one PBC to the other) of particles crossing the periodic boundaries.
In the past, it was an issue since the QCM was not able to take the particles outside of the domain into account.
Some refactor has been done in the previous months and this is not currently an issue anymore, so this "expensive" step was done for nothing.

Solution
Remove this step, meaning the particles crossing the PBC hang outside of the domain, but are still linked to the cell.
Having particles outside of the linked cell is already done for none PBC cells until the cell of "sort_particles_into_subdomain_and_cells".
I tested a few configuration with no issue.
The printing of particle summary is also now done like in DEM to allow printing in multi-core.

Testing
Update the test periodic_boundaries_qcm according to new position, comparison with the particle not on PBCs is still coherent.
Add the test periodic_boundaries_qcm_opposite_flow that is the same test with the same principle, but with the particle on PBCs is placed on the other PBC with an opposite flow. Test is also with 2 cores, so it tests the PBC with CFD-DEM in multi-core.
Tested on the pneumatic conveying example also with no issue.

Former-commit-id: a236260d7a0831a23c0b436f3a7d4e631d52420b